后端 登录注册优化

+++ jinzi-generator-web-backend/src/main/java/com/jinzi/web/common/PageRequest.java
@@ -1,11 +1,13 @@
 package com.jinzi.web.common;
 
+
 import com.jinzi.web.constant.CommonConstant;
 import lombok.Data;
 
 /**
  * 分页请求
  *
+ * @author qimu
  */
 @Data
 public class PageRequest {
@@ -13,12 +15,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private int current = 1;
+    private long current = 1;
 
     /**
      * 页面大小
      */
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     /**
      * 排序字段
